requirements-dev.txt
Authlib==1.6.0
streamlit==1.46.1
tqdm==4.66.4
langchain==0.3.24
langchain-core==0.3.72
langchain-community==0.3.22
langchain-huggingface==0.1.2
langchain-chroma==0.2.4
langchain-ollama==0.3.6
newspaper3k==0.2.8
lxml_html_clean==0.1.1
Unidecode==1.3.8
chromadb==1.0.10
pytest==8.3.2
flashrank==0.2.8
rank_bm25==0.2.2
omegaconf==2.3.0
google-auth==2.36.0
beautifulsoup4==4.12.3
huggingface-hub==0.34.1
google-api-core==2.25.1
google-auth==2.36.0
google-cloud-core==2.4.3
streamlit-id-token-patch @ git+https://github.com/jesseDingley/streamlit-id-token-patch.git@main#egg=streamlit-id-token-patch